Updated legacy image
This image is used in the tutorial. Where it is used, it talks about the "use scene context flag" but the flag is not present in the tutorial image because it is from an older version.
The updated image keeps the same pixel height of the old one so it will not take any additional vertical space in the tutorial.